Align pull command with clone command
