+++ cmd/gh-classroom/pull/pull.go
@@ -1,109 +1,23 @@
 package pull
 
 import (
-	"fmt"
-	"log"
-	"os"
-	"path/filepath"
-	"strings"
-
-	"github.com/MakeNowJust/heredoc"
 	"github.com/cli/cli/v2/pkg/cmdutil"
-	"github.com/cli/go-gh"
-	"github.com/github/gh-classroom/cmd/gh-classroom/shared"
-	"github.com/github/gh-classroom/pkg/classroom"
+	starter_repo "github.com/github/gh-classroom/cmd/gh-classroom/pull/starter-repo"
+	student_repos "github.com/github/gh-classroom/cmd/gh-classroom/pull/student-repos"
 	"github.com/spf13/cobra"
 )
 
 func NewCmdPull(f *cmdutil.Factory) *cobra.Command {
-	var assignmentId int
-	var directory string
-
 	cmd := &cobra.Command{
-		Use:   "pull",
-		Short: "Run pull on an existing set of repositories",
-		Long: heredoc.Doc(`Iterate through an existing set of directories that
-		were previously cloned and run pull to get any new commits.
-		`),
-		Run: func(cmd *cobra.Command, args []string) {
-			client, err := gh.RESTClient(nil)
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			if assignmentId == 0 {
-				classroom, err := shared.PromptForClassroom(client)
-
-				if err != nil {
-					log.Fatal(err)
-				}
-				classroomId := classroom.Id
-
-				assignment, err := shared.PromptForAssignment(client, classroomId)
-				if err != nil {
-					log.Fatal(err)
-				}
-				assignmentId = assignment.Id
-			}
-
-			assignment, err := classroom.GetAssignment(client, assignmentId)
-
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			if strings.HasPrefix(directory, "~") {
-				dirname, _ := os.UserHomeDir()
-				directory = filepath.Join(dirname, directory[1:])
-			}
-
-			fullPath, err := filepath.Abs(filepath.Join(directory, assignment.Slug+"-submissions"))
-
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			if _, err := os.Stat(fullPath); os.IsNotExist(err) {
-				fmt.Println("Directories don't exist run gh classroom clone first")
-				if err != nil {
-					log.Fatal(err)
-				}
-			}
-
-			//Save off the cwd so we can restore when we run gh sync
-			baseDir, err := os.Getwd()
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			entries, err := os.ReadDir(fullPath)
-			if err != nil {
-				log.Fatal(err)
-			}
-
-			for _, r := range entries {
-				clonePath := filepath.Join(fullPath, r.Name())
-				fmt.Printf("Pulling repo: %v\n", clonePath)
-				err = os.Chdir(clonePath)
-				if err != nil {
-					log.Fatal(err)
-				}
-				_, _, err := gh.Exec("repo", "sync")
-				if err != nil {
-					//Don't bail on an error the repo could have changes preventing
-					//a pull, continue with rest of repos
-					fmt.Println(err)
-				}
-				err = os.Chdir(baseDir)
-				if err != nil {
-					log.Fatal(err)
-				}
-			}
-		},
+		Use:     "pull <starter-code|student-repos>",
+		Short:   "Pull starter code or a student's submissions",
+		Example: "",
 	}
 
-	cmd.Flags().IntVarP(&assignmentId, "assignment-id", "a", 0, "ID of the assignment")
-	cmd.Flags().StringVarP(&directory, "directory", "d", ".", "Directory to that the repos live in")
+	cmd.AddCommand(starter_repo.NewCmdStarterRepoPull(f))
+	cmd.AddCommand(student_repos.NewCmdStudentReposPull(f))
 
+	cmd.PersistentFlags().StringP("directory", "d", ".", "Directory to clone into")
+	cmd.PersistentFlags().IntP("assignment-id", "a", 0, "ID of the assignment")
 	return cmd
 }
